feat: add get tickets route
